Implemented the internal VideoPlayer class
It will take care of playing the video and thus computing the current frame in an external thread.
Thus, it removes the need in the user code to update the video before drawing it
Internally, it solves one issue which is the inability to compute this frame in the draw() function since it is marked as const.
Later, both sf::Video and sf::VideoStream will use it internally.